requirements.txt
pyaudio
pyautogui
pyperclip 
openai-whisper
opencv-python
pillow
mediapipe
customtkinter
dragonfly2